backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Mail
python-dotenv
langchain-community
langchain-huggingface
langchain-core
langchain-classic
transformers
torch
faiss-cpu
sentence-transformers